+++ datapkg/cli/download.py
@@ -1,5 +1,4 @@
 import os
-import fnmatch
 
 from datapkg.cli.base import Command
 from datapkg.download import PackageDownloader
@@ -16,11 +15,18 @@ class DownloadCommand(Command):
 Download a package (i.e. metadata and resources) specified by {src-spec} to
 {path}.
 
-Resources to retrieve are selected based on optional match of resource
-format against glob-style format-pattern and url-pattern, e.g.::
+Resources to retrieve are selected interactively if no format-pattern is given.
+If provided, the optional glob-style format-pattern and url-pattern arguments
+are matched against the format and url of the resource to determine whether it
+should be retrieved.
 
-    # download all resources associated with ckan package name
+    # download the package specified by ckan://name to path-on-disk
+    # selecting the resources to retrieve interactively
     download ckan://name path-on-disk
+
+    # download all resources
+    # Note need to quote *
+    download ckan://name path-on-disk "*"
 
     # download only those resources that have format 'csv' (or 'CSV')
     download ckan://name path-on-disk csv
@@ -29,22 +35,25 @@ format against glob-style format-pattern and url-pattern, e.g.::
     download ckan://name path-on-disk xml/*
 
     # download only those resources that have a url starting http://abc (and any format)
-    download ckan://name path-on-disk * http://abc*
+    download ckan://name path-on-disk "*" http://abc*
 '''
 
     def run(self, options, args):
+        pkg_downloader = PackageDownloader(verbose=True)
         spec_from = args[0]
         dest_path_base = args[1]
-        formatpat = args[2] if len(args) > 2 else '*'
-        urlpat = args[3] if len(args) > 3 else '*'
-        def filterfunc(resource, count):
-            match = fnmatch.fnmatch(resource.get('format', ''), formatpat)
-            match = match and fnmatch.fnmatch(resource['url'], urlpat)
-            return match
+        if len(args) > 2:
+            formatpat = args[2]
+            urlpat = args[3] if len(args) > 3 else '*'
+            filterfunc = PackageDownloader.make_glob_filterfunc(
+                    formatpat,
+                    urlpat)
+        else:
+            filterfunc = pkg_downloader.filterfunc_interactive_choice
+
         index, path = self.index_from_spec(spec_from)
         pkg = index.get(path)
         dest_path = os.path.join(dest_path_base, pkg.name)
-        pkg_downloader = PackageDownloader(verbose=True)
         pkg_downloader.download(pkg, dest_path, filterfunc)